services/ocr-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
gunicorn==21.2.0

# OCR引擎
opencv-python==4.6.0.66
Pillow==10.0.1
numpy==1.24.3
pytesseract==0.3.10
paddlepaddle==2.5.2
paddleocr==2.7.3
easyocr==1.7.0

# 数据处理
scikit-image==0.21.0
scipy==1.11.4

# 中文文本处理
jieba==0.42.1

# 缓存和数据库
redis==5.0.1
sqlalchemy==2.0.23

# JWT认证
PyJWT==2.8.0

# 异步任务
celery==5.3.4

# HTTP客户端
requests==2.31.0

# 工具库
python-dotenv==1.0.0
marshmallow==3.20.1
python-multipart==0.0.6
psutil==5.9.6

# 开发工具
pytest==7.4.3
pytest-flask==1.3.0
black==23.9.1
flake8==6.1.0